docs-Make change from session::sc to session::ac slightly more clear
The documentation page for the Session functions include:

"The `session::ac` function replaces the `session::sc` function and returns the current user's access method."

This commit changes this to keep the description of the function in its own sentence, and mention the change in a separate note. This is consistent with how other parts of the documentation flag when a new version of SuurealDB has changed a function name, such as `string::endsWith`.