docs-Fix misspelling of IndexedDB in CLI `surreal start` command's documentation
On the documentation page for the SurrealDB CLI tool's `start` command, the description of the `path` argument says that `indxdb` can be specified to use "IndxDB" as the storage backend. The argument value is correct, but the name of the storage backend is misspelling of "IndexedDB".

This commit changes this misspelling to "IndexedDB".